Fix homepage country flag links
