+++ .eslintrc.json
@@ -24,7 +24,8 @@
     "react/no-array-index-key": 0,
     "react/prefer-stateless-function": 0,
     "max-classes-per-file": 0,
-    "react/jsx-props-no-spreading": 0
+    "react/jsx-props-no-spreading": 0,
+    "jsx-a11y/anchor-is-valid": ["error", { "components": ["Link"], "specialLink": ["hrefLeft", "hrefRight"], "aspects": ["invalidHref", "preferButton"] }]
   },
   "globals": {
     "React": "writable"
