+++ pages/api/countries/[code]/index.js
@@ -1,10 +1,10 @@
 import nextConnect from 'next-connect';
-import CountryService from '../../../services/Country.service';
+import CountryService from '../../../../services/Country.service';
 
 const handler = nextConnect();
 
 handler.get(async (req, res) => {
-  const country = await CountryService.findByCode(req.query.code);
+  const country = await CountryService.findByShortName(req.query.code);
   res.status(200).json(country);
 });
